Support arbitrary length arithmetic expressions
Before this change, you could only supply "offsets" in the form of
BASE+OFF or BASE-OFF where BASE was a label or number and OFF was a
number. After looking at the Merlin manual, the assembly actually
supports 4 arithmetic operations: add, subtract, divide, and multiply.
Evaluation is done left to right and parenthesis are disallowed.

Another constraint is length--"the maximum allowable combined
OPERAND+COMMENT length is 64 characters"--but this is not enforced by
a2asm.